Add created by label during gcp cluster creation
