+++ components/cli/pkg/commands/setup.go
@@ -19,14 +19,8 @@
 package commands
 
 import (
-	"bufio"
-	"encoding/json"
 	"fmt"
-	"math/rand"
 	"os"
-	"os/exec"
-	"strconv"
-	"time"
 
 	"github.com/fatih/color"
 	"github.com/manifoldco/promptui"
@@ -34,17 +28,6 @@ import (
 	"github.com/cellery-io/sdk/components/cli/pkg/constants"
 	"github.com/cellery-io/sdk/components/cli/pkg/util"
 )
-
-var uniqueNumber string
-var projectName string
-var accountName string
-var region string
-var zone string
-
-func init() {
-	rand.Seed(time.Now().UTC().UnixNano())
-	uniqueNumber = strconv.Itoa(rand.Intn(1000))
-}
 
 func RunSetup() {
 	selectTemplate := &promptui.SelectTemplates{
@@ -123,41 +106,4 @@ func selectEnvironment() error {
 	fmt.Println("To create your first project, execute the command: ")
 	fmt.Println("  $ cellery init ")
 	return nil
-}
-
-func getGcpData() (string, string, string, string) {
-	cmd := exec.Command("gcloud", "config", "list", "--format", "json")
-	stdoutReader, _ := cmd.StdoutPipe()
-	stdoutScanner := bufio.NewScanner(stdoutReader)
-	output := ""
-	go func() {
-		for stdoutScanner.Scan() {
-			output = output + stdoutScanner.Text()
-		}
-	}()
-
-	stderrReader, _ := cmd.StderrPipe()
-	stderrScanner := bufio.NewScanner(stderrReader)
-
-	go func() {
-		for stderrScanner.Scan() {
-			fmt.Println(stderrScanner.Text())
-		}
-	}()
-	err := cmd.Start()
-	if err != nil {
-		util.ExitWithErrorMessage("Error occurred while getting gcp data", err)
-	}
-	err = cmd.Wait()
-	if err != nil {
-		util.ExitWithErrorMessage("Error occurred while getting gcp data", err)
-	}
-
-	jsonOutput := &util.Gcp{}
-
-	errJson := json.Unmarshal([]byte(output), jsonOutput)
-	if errJson != nil {
-		fmt.Println(errJson)
-	}
-	return jsonOutput.Core.Project, jsonOutput.Core.Account, jsonOutput.Compute.Region, jsonOutput.Compute.Zone
 }
