+++ components/cli/pkg/util/utils.go
@@ -489,3 +489,12 @@ public function main(string action, cellery:ImageName iName, map<cellery:ImageNa
 
 	return targetFilePath, nil
 }
+
+func ConvertToAlphanumeric(input, replacement string) string {
+	reg, err := regexp.Compile("[^A-Za-z0-9]+")
+	if err != nil {
+		ExitWithErrorMessage("Error making regex", err)
+	}
+	processedString := reg.ReplaceAllString(input, replacement)
+	return processedString
+}
